+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 Currently, this repo is in Prerelease. When it is released, this project will adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 ========
 
+## [Unreleased]
+
+### Changes:
+
+-   Select: selectedOption holds the selected value, instead of just the default.
+
 ## 0.19.1
 
 ### Breaking Changes